Refactor/price to bigdecimal

+++ back-end/src/main/java/org/uorderflow/dto/product/ProductCreateDTO.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record ProductCreateDTO(
         @NotBlank @Size(max = 100) String name,
         @NotBlank @Size(max = 255) String description,
         @NotBlank @Size(max = 2048) String image,
-        @NotNull @Positive Double price,
+        @NotNull @Positive BigDecimal price,
         @NotNull Long productCategoryId,
         @NotNull Boolean isAvailable
 ) {
